tests/submission/Infraware/Forms/contents/Forms/Not Use/input_files.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>input_files</h3>
	</p>
	
	<li>Select a file.</li>
	
	<hr>

	<div id="log"></div> 

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  name="input_form">
	<p><input type='file' id='input_file' onchange='change()'></p>
	</form>

	<script>
		var t1 = async_test('input files test');
		
		function change() {
			var input_file = document.getElementById("input_file");
			var files = input_file.files;

			if (typeof(files) == "object") {
				t1.step(function() {
					assert_equals(files.length, 1, "files attribute is not correct.");
				});
			} else {
				t1.step(function() {
					assert_unreached("files attribute is not exist.");
				});
			}
			t1.done();
		}

	</script>

 </body>
</html>
